Code refactor: edit profile  page component
